HID-1831: re-enable language E2E
